backend/requirements.txt
django>=6.0,<7.0
djangorestframework>=3.17,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.9,<5.0
PyJWT>=2.12,<3.0
